Added notrequests.patch() function to spoof `import requests`.
The idea is you can maybe fake out a lot of packages that rely on
requests but only use some of the API. In particular, craigslist imports
`requests.exceptions` but only uses RequestException.

No doubt there are lots of packages that import different parts of
requests, and it would be nice to fake out all of them to work
perfectly on App Engine. This will do for now.